abstract out generic plugin main function
